Fixed onResume issue - removed it
It seems more prudent to leave the option
and timing of a refresh up to an informed
user - rather than try to do it
automatically.  This way the course list
can stay more stable, and the whole app
is better for it.  Perhaps when complexity
is added later it may become prudent, but
for now a big button will suffice.

  FIXED: onResume mutiplication issue
  FIXED: UI consistency in appearance