UI update, last page to video icons
Trying to make the UI prettier, added
youtube video icons to the nugget page.
The icons are not intelligently cached
at the moment, so the scroll is a bit
jerky.  The CourseRev should probably
be commited to a SQL database at this
point, and a cursor adapter used for
the management of the listviews rather
than the custom adapters.

  ADDED: Icons to nuggets